refactor: TimeList 컴포넌트 네이밍 수정 - 페이지와 네이밍 통일
